make includes consistent
