+++ src/cast_core/CMakeLists.txt
@@ -12,10 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-project(product-perf VERSION 0.0.1 LANGUAGES CXX)
+project(stm VERSION 0.0.1 LANGUAGES CXX)
 
 CreateGennyTargets(
-    NAME    product-perf
+    NAME    cast_core
     TYPE    SHARED
     DEPENDS
         gennylib
