+++ src/cast_core/include/cast_core/actors/CrudActor.hpp
@@ -10,10 +10,9 @@
 #include <gennylib/ExecutionStrategy.hpp>
 #include <gennylib/PhaseLoop.hpp>
 #include <gennylib/context.hpp>
+
 #include <value_generators/DefaultRandom.hpp>
 #include <value_generators/value_generators.hpp>
-
-#include <stm/config/RunCommandConfig.hpp>
 
 namespace genny::actor {
 
@@ -67,6 +66,8 @@ namespace genny::actor {
  *     Operation:
  *       OperationName: drop
  * ```
+ *
+ * Owner: STM
  */
 class CrudActor : public Actor {
