+++ src/cast_core/src/actors/HelloWorld.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <stm/actors/HelloWorld.hpp>
+#include <cast_core/actors/HelloWorld.hpp>
 
 #include <string>
