+++ src/cast_core/test/RunCommandActor_test.cpp
@@ -19,15 +19,15 @@
 #include <bsoncxx/builder/stream/document.hpp>
 #include <bsoncxx/json.hpp>
 
-#include <stm/actors/RunCommand.hpp>
+#include <mongocxx/read_preference.hpp>
+
+#include <cast_core/actors/RunCommand.hpp>
 
 #include <gennylib/MongoException.hpp>
+
 #include <testlib/ActorHelper.hpp>
 #include <testlib/MongoTestFixture.hpp>
-
 #include <testlib/helpers.hpp>
-
-#include <mongocxx/read_preference.hpp>
 
 namespace genny {
 namespace {
