+++ src/gennylib/benchmark/RateLimiter_benchmark.cpp
@@ -21,6 +21,7 @@
 #include <gennylib/config/RateLimiterOptions.hpp>
 #include <gennylib/conventions.hpp>
 #include <gennylib/v1/RateLimiter.hpp>
+
 #include <testlib/ActorHelper.hpp>
 #include <testlib/helpers.hpp>
