+++ src/gennylib/include/gennylib/context.hpp
@@ -35,7 +35,9 @@
 #include <gennylib/Orchestrator.hpp>
 #include <gennylib/conventions.hpp>
 #include <gennylib/v1/ConfigNode.hpp>
+
 #include <metrics/metrics.hpp>
+
 #include <value_generators/DefaultRandom.hpp>
 
 /**
