+++ src/gennylib/src/Actor.cpp
@@ -19,6 +19,5 @@
 #include <gennylib/context.hpp>
 
 namespace genny {
-
 Actor::Actor(ActorContext& context) : _id{context.workload().nextActorId()} {}
 }  // namespace genny
