+++ src/gennylib/src/Orchestrator.cpp
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <gennylib/Orchestrator.hpp>
+
 #include <algorithm>  // std::max
 #include <cassert>
 #include <iostream>
-
-#include <gennylib/Orchestrator.hpp>
 
 namespace {
