--- src/gennylib/src/log.hh
@@ -1,6 +1,0 @@
-#ifndef HEADER_AC9B760C_005E_42BD_8526_16EC12A9E0B2_INCLUDED
-#define HEADER_AC9B760C_005E_42BD_8526_16EC12A9E0B2_INCLUDED
-
-#include <boost/log/trivial.hpp>
-
-#endif  // HEADER_AC9B760C_005E_42BD_8526_16EC12A9E0B2_INCLUDED
