+++ src/scale/include/scale/actors/Loader.hpp
@@ -19,7 +19,9 @@
 #include <gennylib/Actor.hpp>
 #include <gennylib/PhaseLoop.hpp>
 #include <gennylib/context.hpp>
+
 #include <metrics/metrics.hpp>
+
 #include <value_generators/value_generators.hpp>
 
 namespace genny::actor {
