+++ src/scale/src/MultiCollectionQuery.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <product-perf/actors/MultiCollectionUpdate.hpp>
+#include <scale/actors/MultiCollectionQuery.hpp>
 
 #include <chrono>
 #include <memory>
@@ -20,64 +20,70 @@
 #include <thread>
 
 #include <bsoncxx/json.hpp>
+
 #include <mongocxx/client.hpp>
 #include <mongocxx/pool.hpp>
+
 #include <yaml-cpp/yaml.h>
 
 #include <boost/log/trivial.hpp>
 
 #include <gennylib/Cast.hpp>
 #include <gennylib/context.hpp>
+
 #include <value_generators/value_generators.hpp>
 
 namespace genny::actor {
 
 /** @private */
-struct MultiCollectionUpdate::PhaseConfig {
+struct MultiCollectionQuery::PhaseConfig {
     PhaseConfig(PhaseContext& context, genny::DefaultRandom& rng, mongocxx::pool::entry& client)
         : database{(*client)[context.get<std::string>("Database")]},
           numCollections{context.get<UIntSpec, true>("CollectionCount")},
-          queryDocument{value_generators::makeDoc(context.get("UpdateFilter"), rng)},
-          updateDocument{value_generators::makeDoc(context.get("Update"), rng)},
+          filterDocument{value_generators::makeDoc(context.get("Filter"), rng)},
           uniformDistribution{0, numCollections},
-          minDelay{context.get<TimeSpec, false>("MinDelay").value_or(TimeSpec(0))} {}
+          minDelay{context.get<TimeSpec, false>("MinDelay").value_or(genny::TimeSpec(0))} {}
 
     mongocxx::database database;
     size_t numCollections;
-    std::unique_ptr<value_generators::DocumentGenerator> queryDocument;
-    std::unique_ptr<value_generators::DocumentGenerator> updateDocument;
-    // TODO: Enable passing in update options.
-    //    std::unique_ptr<value_generators::DocumentGenerator>  updateOptions;
-
+    std::unique_ptr<value_generators::DocumentGenerator> filterDocument;
     // uniform distribution random int for selecting collection
     std::uniform_int_distribution<size_t> uniformDistribution;
     std::chrono::milliseconds minDelay;
+    mongocxx::options::find options;
 };
 
-void MultiCollectionUpdate::run() {
+void MultiCollectionQuery::run() {
     for (auto&& config : _loop) {
         for (auto&& _ : config) {
             // Take a timestamp -- remove after TIG-1155
             auto startTime = std::chrono::steady_clock::now();
 
             // Select a collection
+            // This area is ripe for defining a collection generator, based off a string generator.
+            // It could look like: collection: {^Concat: [Collection, ^RandomInt: {min: 0, max:
+            // *CollectionCount]} Requires a string concat generator, and a translation of a string
+            // to a collection
             auto collectionNumber = config->uniformDistribution(_rng);
             auto collectionName = "Collection" + std::to_string(collectionNumber);
             auto collection = config->database[collectionName];
 
-            // Perform update
-            bsoncxx::builder::stream::document updateFilter{};
-            bsoncxx::builder::stream::document update{};
-            auto filterView = config->queryDocument->view(updateFilter);
-            auto updateView = config->updateDocument->view(update);
+            // Perform a query
+            bsoncxx::builder::stream::document filter{};
+            auto filterView = config->filterDocument->view(filter);
             // BOOST_LOG_TRIVIAL(info) << "Filter is " <<  bsoncxx::to_json(filterView);
-            // BOOST_LOG_TRIVIAL(info) << "Update is " << bsoncxx::to_json(updateView);
             // BOOST_LOG_TRIVIAL(info) << "Collection Name is " << collectionName;
             {
                 // Only time the actual update, not the setup of arguments
-                auto op = _updateTimer.raii();
-                auto result = collection.update_many(filterView, updateView);
-                _updateCount.incr(result->modified_count());
+                auto op = _queryTimer.raii();
+                auto cursor = collection.find(filterView, config->options);
+                // exhaust the cursor
+                uint count = 0;
+                for (auto&& doc : cursor) {
+                    doc.length();
+                    count++;
+                }
+                _documentCount.incr(count);
             }
             // make sure enough time has passed. Sleep if needed -- remove after TIG-1155
             auto elapsedTime = std::chrono::steady_clock::now() - startTime;
@@ -87,16 +93,16 @@ void MultiCollectionUpdate::run() {
     }
 }
 
-MultiCollectionUpdate::MultiCollectionUpdate(genny::ActorContext& context)
+MultiCollectionQuery::MultiCollectionQuery(genny::ActorContext& context)
     : Actor(context),
       _rng{context.workload().createRNG()},
-      _updateTimer{context.timer("updateTime", MultiCollectionUpdate::id())},
-      _updateCount{context.counter("updatedDocuments", MultiCollectionUpdate::id())},
+      _queryTimer{context.timer("queryTime", MultiCollectionQuery::id())},
+      _documentCount{context.counter("returnedDocuments", MultiCollectionQuery::id())},
       _client{std::move(context.client())},
       _loop{context, _rng, _client} {}
 
 namespace {
-auto registerMultiCollectionUpdate =
-    genny::Cast::registerDefault<genny::actor::MultiCollectionUpdate>();
+auto registerMultiCollectionQuery =
+    genny::Cast::registerDefault<genny::actor::MultiCollectionQuery>();
 }
 }  // namespace genny::actor
