+++ src/scale/workloads/CollScan.yml
