+++ src/testlib/src/ActorHelper.cpp
@@ -19,6 +19,7 @@
 
 #include <gennylib/Orchestrator.hpp>
 #include <gennylib/context.hpp>
+
 #include <metrics/MetricsReporter.hpp>
 #include <metrics/metrics.hpp>
