+++ src/testlib/test/ActorHelper_test.cpp
@@ -20,8 +20,8 @@
 #include <gennylib/ActorProducer.hpp>
 #include <gennylib/InvalidConfigurationException.hpp>
 #include <gennylib/context.hpp>
-#include <testlib/ActorHelper.hpp>
 
+#include <testlib/ActorHelper.hpp>
 #include <testlib/helpers.hpp>
 
 using Catch::Matchers::Matches;
