+++ src/value_generators/include/value_generators/value_generators.hpp
@@ -15,10 +15,11 @@
 #ifndef HEADER_E6E05F14_BE21_4A9B_822D_FFD669CFB1B4_INCLUDED
 #define HEADER_E6E05F14_BE21_4A9B_822D_FFD669CFB1B4_INCLUDED
 
-#include <bsoncxx/builder/stream/document.hpp>
 #include <exception>
 #include <optional>
 #include <unordered_map>
+
+#include <bsoncxx/builder/stream/document.hpp>
 
 #include <yaml-cpp/yaml.h>
